--- Projects/project3/.env
@@ -1,3 +1,0 @@
-#These are free API tokens.
-CAT_BREED_TOKEN=mG80UmvuIidsp3ZPi86GIw==ilugQMghVMf4Z3hT
-CAT_IMAGE_TOKEN=live_xze8YHkSkKWaoGB7xOwhaqKyem7cVgOIvl2fiT8zhb40FKk5FlGpuYTABVHLYt7U
