+++ Projects/project6_answer.py
@@ -1,4 +1,4 @@
-# Python Classes - Project 5
+# Python Classes - Project 6
 # Author: Phil van der Linden
 # Date: 05/28/2024
