+++ Projects/project6_rubric.py
@@ -1,0 +1,35 @@
+# Project 6 (100 Points) - Python Inheritance Types
+
+# The goal of this project is to learn about the five different types of Python inheritance.
+
+# Start by reading the following article:
+# https://www.geeksforgeeks.org/types-of-inheritance-python/
+
+# Your task is to create several classes.
+# These classes can either be all in your main file, or seperated into multiple files (bonus points).
+# Next you will use your classes to demonstrate each type of inheritance.
+
+# Scenario:
+# A new database system is being set up at a university you work at.
+# The database will use Python code to determine classes and create different object types.
+
+# Classes:
+# 1. An affiliate class.
+# 2. A student class.
+# 3. An employee class.
+# 4. A professor class.
+# 5. An adjunct class.
+# 6. A student_worker class.
+# 7. An alumni class.
+
+# Inheritance:
+# a. The student class demonstrates single inheritance from the affiliate class.
+# b. The student_worker class demonstrates multiple inheritance from the employee class and student class.
+# c. The adjunct class demonstrates multilevel inheritance from the affiliate, employee, and projessor classes.
+# d. The affiliate class demonstrates hierarchical inheritance with students, employees, and alumni classes.
+# d. All of the classes together will demonstrate hybrid inheritance.
+
+# Other requirements:
+# i. At least one object should be created for each class.
+# ii. Each class should have at least one property.
+# iii. Create two functions. These can be in one or more classes.
