+++ README.md
@@ -1,12 +1,15 @@
 # PythonLearning
 ##### Learn or practice Python programming with this repo!
-_...This course assumes that you know computer and Internet basics, such as using a virtual machine, searching the Internet, and cloning a GitHub repo._
+> _...This course assumes that you have intermediate computer proficiency, such as setting up a virtual machine, searching the Internet, and cloning a GitHub repo._
 
 ## Setup
 
 | Install | Link |
 | ------ | ------ |
-| A Linux Flavor | https://ubuntu.com/download/desktop |
-| Virtual Box | https://www.virtualbox.org/wiki/Downloads |
-| VS Code | https://code.visualstudio.com/download |
-| Git Clone | https://github.com/brok3nwir3/PythonLearning.git |
+| Ubuntu (Linux) | https://ubuntu.com/download/desktop |
+| Virtual Box (VM) | https://www.virtualbox.org/wiki/Downloads |
+| VS Code (IDE) | https://code.visualstudio.com/download |
+| Git Clone (Repo) | https://github.com/brok3nwir3/PythonLearning.git |
+
+## Usage
+Open the <ins>Assignments</ins> folder and start answering questions (beginning at week 1). Use the <ins>Answer Keys</ins> folder to compare your answers against the official answers. Check the bottom of each assignment for homework or project prompts (see the <ins>Projects</ins> folder). Tests and quizzes are assigned every few weeks, and can be located in the <ins>Quizzes & Tests folder</ins>.
